修復：修正因 dragThreshold 與 FragIndex 型別變更導致的測試失敗
- gesture_test.go: 座標差值改為小於新 dragThreshold(0.005)
- packet_test.go: FragIndex 比對型別改為 uint16

+++ server/internal/stream/packet_test.go
@@ -48,7 +48,7 @@ func TestSplitNALUnits(t *testing.T) {
 
 	// Check fragment indices
 	for i, pkt := range packets {
-		if pkt.Header.FragIndex != uint8(i) {
+		if pkt.Header.FragIndex != uint16(i) {
 			t.Errorf("packet %d: fragIndex got %d, want %d", i, pkt.Header.FragIndex, i)
 		}
 		if pkt.Header.FragTotal != 3 {
